Coding standardei tehty 5.1 c:hen

+++ optical_tachometer_FINAL.X/update_lcd.c
@@ -15,11 +15,10 @@
 #include <util/delay.h>
 #include <string.h>
 
-void update_lcd(int rpm)
-
+void update_lcd(uint16_t rpm)
 {
     // Placeholder for calculated rpm
-    int rpmOriginal = rpm;
+    uint16_t rpmOriginal = rpm;
 
     // This will be the number of elements in an array, initialized as 0 in the
     // beginning
